+++ TheWoodlandFamily/DataLibrary/Entities/Room.cs
@@ -8,7 +8,7 @@ namespace EFDataAccessLibrary.Entities
     {
         public int Id { get; set; }
         public string WordKey { get; set; }
-        public int PlayerNumber { get; set; }
+        public byte PlayerNumber { get; set; }
         public List<Player> Players { get; set; } = new List<Player>();
         public List<Card> Deck { get; set; } = new List<Card>();
     }
